requirements.txt
dnspython==2.7.0
loguru==0.7.2
pyaes==1.6.1
motor==3.3.2
pymongo==4.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.0
TgCrypto==1.2.5
uvloop==0.21.0
